Sweep2Wake: Implemented for atmel qt602240_ts
I have adapted this feature it based on the work of thachtunganh,
optimized and revised by some of commit of oliverclaude.

I had to activate the gpio interrupt input mode for the keypad,
researching for the address gpio, of the power button not supplied by samsung sources,
i also had to include some systems of workaround/hack to coexist the two modes of management of the keypad.
Lastly, to include managing switching on/off of the LEDs touchkey from S2W,
i added some changes on resume of cypress-touchkey module.

Sources:
https://github.com/oliverclaude/shooteru-ics-crc-3.0.16-e733189
https://github.com/KangBangKreations/KangBanged-7x30